pyproject.toml
[project]
name = "drf"
version = "0.1.0"
description = ""
authors = [
    {name = "Dmitry Pikov",email = "dmitrypikov@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "psycopg (>=3.2.10,<4.0.0)",
    "psycopg-binary (>=3.2.10,<4.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "django-filter (>=25.1,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "ipython (>=9.6.0,<10.0.0)",
    "coverage (>=7.10.7,<8.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "stripe (>=13.0.1,<14.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=7.0.0,<8.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

